+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-## Coming Soon
+## v1.6.0
 
 ### fatcher
 
--   🚀 add options `validateCode` for passing custom response status code.
--   ⚡️ `normalizeURL` downgrade to use `match`.
+-   🚀 add options `validateCode` for passing custom response status code. ([#174](https://github.com/fatcherjs/fatcher/pull/174))
+-   ⚡️ `normalizeURL` downgrade to use `match`. ([#179](https://github.com/fatcherjs/fatcher/pull/179))
 
 ## v1.5.0